[RD] Build init func's.
This commit adds an interface to all dataset ops to hook a variety of
functionality. One bit of functionality is a helper function to
build the relevant iterator type (using LLVM types). The second uses
a builder to insert LLVM instructions to initialize memory for the
iterator state.

Future work will add an additional interface method to build the
iterator get_next implementation.