Explore using multiclass & defm to simplify lowering.
The input program's dataset ops combine both initialization and dataflow
in a convenient form. To lower to iterator form, we need to tease apart
which parts are for initialization and which are used during actual iteration.

This change defines a `multiclass` that allows us to define 3 interrelated ops:
 1. The baseline op which is used to describe the input program.
 2. The init op, which initializes the iterator state.
 3. The next op, which advances the iterator by one and returns the next value.

The goal of defining these interrelated ops is to explore an implementation of
the lowerings to LLVM IR that is less monolithic (and thus easier to test and
develop) than the previous design.

Note: this change simply changes the tblgen, but has not yet defined appropriate
interfaces or generic passes to support this lowering.

Finally, I'm playihng fast-and-loose with the !rd.Dataset type, as it takes the
places of values (and boolean sentinals) in the `@next` functions, while
describing the aggregate computation / collection in the `@definition` func.